refactor: Token info
Refactored rendering of token info by moving to child component, also renamed props on chronology field component
